requirements.txt
asgiref==3.6.0
attrs==23.1.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
distlib==0.3.6
Django==4.1.6
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
filelock==3.9.0
heroku==0.1.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.2
mysqlclient==2.1.1
packaging==23.1
parsel==1.8.1
Pillow==9.4.0
platformdirs==2.6.2
Protego==0.2.1
psycopg2==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.1.1
python-dateutil==1.5
pytz==2022.7.1
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
Scrapy==2.8.0
service-identity==21.1.0
six==1.16.0
SpeechRecognition==3.9.0
sqlparse==0.4.3
tldextract==3.4.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.18.0
w3lib==2.1.1
zope.interface==6.0
